still refactoring Entrance
