+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "budget-manager",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "private": true,
   "homepage": "http://arcanrun.github.io/budget-manager-dev",
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.13.0",
-    "@vkontakte/icons": "^1.20.0",
+    "@vkontakte/icons": "^1.20.2",
     "@vkontakte/mvk-mini-apps-scroll-helper": "^1.2.0",
     "@vkontakte/vk-connect": "^1.9.1",
     "@vkontakte/vkui": "^2.34.1",
@@ -14,13 +14,13 @@
     "currency-symbol-map": "^4.0.4",
     "deepcopy": "^2.0.0",
     "flow-bin": "^0.95.1",
-    "moment": "^2.24.0",
-    "node-sass": "^4.13.1",
+    "moment": "^2.25.2",
+    "node-sass": "^4.14.0",
     "numeral": "^2.0.6",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-chartist": "^0.13.3",
-    "react-day-picker": "^7.4.0",
+    "react-day-picker": "^7.4.8",
     "react-dom": "^16.13.1",
     "react-id-swiper": "^2.4.0",
     "react-joyride": "^2.2.1",
