+++ src/components/Entrance/SubComponents/EntranceSwiper/EntranceSwiper.module.css
@@ -1,0 +1,9 @@
+.item {
+  display: flex;
+  flex-flow: column wrap;
+  justify-content: center;
+  align-items: center;
+  font-weight: 200;
+  font-size: 20px;
+  text-align: center;
+}
