+++ src/components/Entrance/SubComponents/EntranceSwiper/costumizedSwiper.css
