+++ src/components/Entrance/SubComponents/EntranceSwiper/helpers.js
